Implement reason for admin rejected bookings
Support for reasons for rejected/revoked bookings done by admin, as
the reason is required when those actions are done by admins.

+++ backend/internal/handlers/booking.go
@@ -92,7 +92,13 @@ func (h *BookingHandler) DeleteBookingByID(w http.ResponseWriter, r *http.Reques
 		return
 	}
 
-	userID := r.Context().Value(mw.ContextKeyUserID).(string)
+	userID, ok := r.Context().Value(mw.ContextKeyUserID).(string)
+	if !ok || userID == "" {
+		// TODO: error handling and logging
+		log.Println("User ID not found in context, interrupting...")
+		http.Error(w, "Unauthorized: User ID not found", http.StatusUnauthorized)
+		return
+	}
 
 	err = h.service.DeleteBookingByID(bookingID, userID)
 	if err != nil {
